requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-groq==0.1.5
python-dotenv==1.0.1